+++ stripe-core/src/main/java/com/stripe/android/core/networking/SendAnalyticsRequestV2Worker.kt
@@ -1,5 +1,6 @@
 package com.stripe.android.core.networking
 
+import android.app.Application
 import android.content.Context
 import androidx.annotation.VisibleForTesting
 import androidx.work.CoroutineWorker
@@ -7,8 +8,6 @@ import androidx.work.Data
 import androidx.work.WorkerParameters
 import androidx.work.workDataOf
 import com.stripe.android.core.exception.InvalidRequestException
-import kotlinx.serialization.encodeToString
-import kotlinx.serialization.json.Json
 
 private const val DataKey = "data"
 private const val MaxAttempts = 5
@@ -26,7 +25,7 @@ internal class SendAnalyticsRequestV2Worker(
                 Result.success()
             },
             onFailure = { error ->
-                if (error.shouldRetry && runAttemptCount < MaxAttempts) {
+                if (error.shouldRetry && runAttemptCount < MaxAttempts - 1) {
                     Result.retry()
                 } else {
                     Result.failure()
@@ -35,9 +34,18 @@ internal class SendAnalyticsRequestV2Worker(
         )
     }
 
-    private inline fun withRequest(block: (AnalyticsRequestV2) -> Result): Result {
-        val request = getRequest(inputData) ?: return Result.failure()
+    private suspend inline fun withRequest(block: (AnalyticsRequestV2) -> Result): Result {
+        val id = inputData.getString(DataKey) ?: return Result.failure()
+        val request = storage(applicationContext).retrieve(id) ?: return Result.failure()
         val workManagerRequest = request.withWorkManagerParams(runAttemptCount)
+
+        val result = block(workManagerRequest)
+        val canRemove = result != Result.retry()
+
+        if (canRemove) {
+            storage(applicationContext).delete(id)
+        }
+
         return block(workManagerRequest)
     }
 
@@ -48,23 +56,22 @@ internal class SendAnalyticsRequestV2Worker(
         var networkClient: StripeNetworkClient = DefaultStripeNetworkClient()
             private set
 
-        fun createInputData(request: AnalyticsRequestV2): Data {
-            val encodedRequest = Json.encodeToString(request)
-            return workDataOf(DataKey to encodedRequest)
-        }
+        var storage: (Context) -> AnalyticsRequestV2Storage =
+            { RealAnalyticsRequestV2Storage(it.applicationContext as Application) }
+            private set
 
-        private fun getRequest(data: Data): AnalyticsRequestV2? {
-            val encodedRequest = data.getString(DataKey)
-            return encodedRequest?.let {
-                runCatching<AnalyticsRequestV2> {
-                    Json.decodeFromString(it)
-                }.getOrNull()
-            }
+        fun createInputData(id: String): Data {
+            return workDataOf(DataKey to id)
         }
 
         @VisibleForTesting
         fun setNetworkClient(networkClient: StripeNetworkClient) {
             this.networkClient = networkClient
+        }
+
+        @VisibleForTesting
+        fun setStorage(storage: AnalyticsRequestV2Storage) {
+            this.storage = { storage }
         }
     }
 }